backend/requirements.txt
google-cloud-aiplatform>=1.64.0
google-cloud-bigquery>=3.13.0
google-cloud-secret-manager>=2.10.0
google-cloud-scheduler==2.16.1
mysql-connector-python==9.4.0
pandas>=2.0.0
rich>=13.0.0
uvicorn==0.35.0
fastapi==0.116.1
dotenv==0.9.9
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
google-cloud-scheduler
alembic
